move d.ts to separate folder
