--- packages/ace-linters/index.d.ts
@@ -1,9 +1,0 @@
-import {AceLinters} from "./services/language-service";
-
-export {AceLinters} from "./services/language-service";
-export {LanguageProvider} from "./language-provider";
-export {MessageController} from "./message-controller";
-export * from "./type-converters/converters";
-
-export function registerStyles(): void;
-
