+++ packages/ace-linters/message-types.ts
@@ -1,6 +1,6 @@
 import {Ace} from "ace-code";
 import {FormattingOptions} from "vscode-languageserver-types";
-import {AceLinters} from "./services/language-service";
+import {AceLinters} from "./types";
 import ServiceOptions = AceLinters.ServiceOptions;
 
 export abstract class BaseMessage {
