+++ packages/ace-linters/services/base-service.ts
@@ -1,6 +1,6 @@
 import {FormattingOptions} from "vscode-languageserver-types";
 import {Ace} from "ace-code";
-import {AceLinters} from "./language-service";
+import {AceLinters} from "../types";
 
 export abstract class BaseService<OptionsType extends AceLinters.ServiceOptions = AceLinters.ServiceOptions> implements AceLinters.LanguageService {
     abstract $service;
