+++ packages/ace-linters/services/html/html-service.ts
@@ -7,7 +7,7 @@ import {
     toCompletions,
     toResolvedCompletion,
     toTooltip
-} from "../../type-converters/vscode-converters";
+} from "../../type-converters/lsp-converters";
 import {HTMLFormatConfiguration} from "vscode-html-languageservice/lib/umd/htmlLanguageTypes";
 import {BaseService} from "../base-service";
