+++ packages/ace-linters/types/index.d.ts
@@ -1,0 +1,9 @@
+import {AceLinters} from "./language-service";
+
+export {AceLinters} from "./language-service";
+export {LanguageProvider} from "../language-provider";
+export {MessageController} from "../message-controller";
+export * from "./converters";
+
+export function registerStyles(): void;
+
