--- packages/ace-linters/worker-loader.d.ts
@@ -1,7 +1,0 @@
-declare module "*.worker.ts" {
-    class WebpackWorker extends Worker {
-        constructor();
-    }
-    export default WebpackWorker
-}
-
